+++ tobkiri_runtime/ecosystem/defaultspack/webapp/src/App.tsx
@@ -15,6 +15,7 @@ import { AuthorityApprovalNotice } from "./components/AuthorityApprovalNotice";
 import { AuthorityApprovalWindow } from "./components/AuthorityApprovalWindow";
 import { ApprovalDecisionSurface } from "./components/ApprovalDecisionSurface";
 import { ErrorNotice } from "./components/ErrorNotice";
+import { CLOSE_RIGHT_SIDEBAR_PANEL_REQUEST } from "./components/RightSidebar";
 import { CodingCockpit } from "./components/coding/CodingCockpit";
 import { HostPermissionsPage } from "./hostPermissions/HostPermissionsPage";
 import { ConversationSpotlight } from "./components/ConversationSpotlight";
@@ -6611,6 +6612,8 @@ export function ChatApp() {
       }
       return;
     }
+    setActiveSidebarItemId(CLOSE_RIGHT_SIDEBAR_PANEL_REQUEST);
+    setSidebarSelectionTick((value) => value + 1);
     setIsGenerating(true);
     cancelPendingMentionAttachments();
     setError(null);